Move parallel_safety to safety folder
